🏗 Fetch and parse the listings file
🚧 Fetch the list from file for now (easier to test since it does not make an API call for every reload).

The result is saved in a transient property which takes care of caching -- only fetch new content if the cache time has expired. This may be a usable solution or then it may need to be fully replaced by the import functionality.

Use the ListParser to parse the JSON into proper listing values.