🏗 Add a configuration class for the plugin
The class will take care of defining the hard-coded options and the db config keys,
and has helper methods to fetch pand parse options values.

The final URL will probably be hard-coded, with just a login key saved in the
options but for now, and for testing, the entire URL is saved in the options.