🚧 Add filtering by apartment type
This is still just a crude demo for filtering, to be refactored into
something more coherent.

The apartment type filter only shows types that are present in the
list.

Filtering by a field type is only done if some values have been selected
-- this allows showing all by default but also just selecting the types
you want to see instead of having to deselect all you want to filter out.