🥅 Improve error handling & caching of listing API result
If the API does not return valid XML data, do not cache the result in
transients and throw an exception. The UI will display an error message.
Log the API result value in the error log.

This happens for example if the shortcode is inserted on a page before
the user accout is set in plugin settings (or if the account name is
incorrect). In this case, after fixing the plugin settings, reloading
the page will re-fetch the data from the API.